softdesk/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.11.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
install==1.3.5
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
self==2020.12.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
